% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Operator potoku (pipe)}
\usage{
lhs \%>\% rhs
}
\value{
wynik dzialania \code{rhs(lhs)}
}
\description{
Re-eksport \code{\%>\%} z pakietu magrittr, zeby dzialal wewnatrz pakietu.
To rozwiazuje blad "could not find function %>%".
}
\keyword{internal}
