% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_sales_ts}
\alias{validate_sales_ts}
\title{Validate sales time series}
\usage{
validate_sales_ts(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
list with quality report
}
\description{
Validate sales time series
}
