% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_sales_metrics}
\alias{compute_sales_metrics}
\title{Compute sales metrics}
\usage{
compute_sales_metrics(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
tibble
}
\description{
Compute sales metrics
}
