% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{create_management_summary}
\alias{create_management_summary}
\title{Create management summary}
\usage{
create_management_summary(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
list
}
\description{
Create management summary
}
