% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{sales_ts_logic}
\alias{sales_ts_logic}
\title{Sales analysis pipeline}
\usage{
sales_ts_logic(df, store = NULL, family = NULL)
}
\arguments{
\item{df}{dataframe}

\item{store}{optional store id}

\item{family}{optional product category}
}
\value{
list
}
\description{
Sales analysis pipeline
}
